Move code related to generating portindices to a package portindex1.0, encapsulate old Tcl implementation, provide SQLite-based PortIndex implementation
The PortIndex is currently write-only. SQLite index does not support archive
mode (do we even use that anymore?).

Code querying the PortIndex should (and will) be moved to portindex1.0, too.

Some numbers: When generating the PortIndex from scratch, the runtime is
dominated by the evaluation time of the Portfiles. Both implementations achieve
similar runtime. When few ports need to be updated, SQLite-based is up to 25%
faster.

git-svn-id: https://svn.macports.org/repository/macports/users/cal@103014 d073be05-634f-4543-b044-5fe20cf6d1d6