rename private branch to better represent its purpose from now on
I've noticed that most of the things I wanted to speed up and improve
using the SQLite PortIndex are actually implemented in the port client
(i.e., the script you run from the command line).
 - evaluating pseudo-ports (base behind port doesn't know what
   "outdated" is, which makes it unnecessarily difficult to write other
   clients, like a GUI, for MacPorts).
 - evaluating expressions selecting port lists (i.e., maintainer:^cal
   and name:p5-*)
Those were some of the places I would have to touch to get the SQLite
PortIndex integrated, among a lot of others somewhere in macports1.0.

This branch will serve as a place to separate the port client and the
macports API, which have become very entangled over the years. I'm
trying to implement this by rewriting macports1.0 (as macports2.0) and
will finish by re-implementing the port client against this new API.

While I'm at this I'll make sure to document the API (so others could
use it, to e.g., write a GUI).

If you want to help out, feel free to commit to this branch, even though
it is in my user directory.

git-svn-id: https://svn.macports.org/repository/macports/users/cal@103639 d073be05-634f-4543-b044-5fe20cf6d1d6